commit 4
manage to fix featured section and finish ed it off, newsletter section, arrivals section is done also. Now currently working on the deal section.

Notice the icon section is a little bit off to the right side when using screen with bigger res than 1024px, think we should put the name of the  book in it's box header so it will be easier for customer to read. Will try to fix these later.