commit 5
finished the website, good job me.

Still there are some aspect need to adjust and add more color to the website so it doesn't look to simple, Also need to start linking all the page together, still I think we need some a contact page, just make it simple, book detail page (just make for one divide section, and copy pass them), categories maybe it a good idea to make one, it can be simple. BTW you can copy the header and footer in to the other pages as well.