commit 7
finished in completing the home page though the name and color may need to change.

the collection(category) is made and it already have the featured, new arrivals including it banner but still I think we need to add more section for variety.

the book detail page have not been made, consider doing this in the next commit.